added working thrust benchmark

+++ test/integ/reduce/src/Reduce.cpp
@@ -136,7 +136,7 @@ TEST_CASE("Test reduce", "[reduce]")
         // test against thrust
         thrust::device_vector<std::uint64_t> deviceReduce(reduce);
         start = std::chrono::high_resolution_clock::now();
-        tSum = thrust::reduce(deviceReduce.begin(), deviceReduce.end(), 0, thrust::plus<std::uint64_t>());
+        tSum = thrust::reduce(deviceReduce.begin(), deviceReduce.end(), static_cast<std::uint64_t>(0), thrust::plus<std::uint64_t>());
         end = std::chrono::high_resolution_clock::now();
         std::cout << "Runtime of thrust reduce: ";
         std::cout << std::chrono::duration_cast<std::chrono::microseconds>(end - start).count() << " microseconds\n";
